kafka_video_streaming/requirements.txt
kafka-python
opencv-python
numpy


